Updating policy files to reflect the current Duo Desktop policy structure (#267)
* Update policies.py to use Duo Desktop section

The Duo Desktop section replaced the device_health_app section in the official admin API docs, but this example was never updated to reflect that. This commit updates this to use the correct section.

* Update policies_advanced.py to use Duo Desktop

The Duo Desktop section replaced the device_health_app section in the official admin API docs, but this example was never updated to reflect that. This commit updates this to use the correct section.

* Remove the prompt_to_install, its not a param anymore

* Ensuring the prompt_to_install param isn't there, as that got removed

* Ensuring the old device_health_app function is still there for backwards compatibility

* Ensuring device_health_app function is maintained for backwards compatibility